+++ DropBGBackend/src/main/java/com/example/DropBGBackend/Service/RazorpayService.java
@@ -6,6 +6,6 @@ import com.razorpay.RazorpayException;
 import java.util.Map;
 
 public interface RazorpayService {
-    Order createdOrder(Double amount, String currency) throws RazorpayException;
+    Order createdOrder(Integer amount, String currency) throws RazorpayException;
     Map<String,Object> verifyPayment(String paymentId) throws RazorpayException;
 }
